refactor: move URL params logic to URLParams class and use urlsafe base64
