+++ src/pumpwood_streamlit.egg-info/requires.txt
@@ -1,9 +1,9 @@
-extra-streamlit-components
-loguru>=0.7.3
-lxml
-openpyxl
-pandas
-pumpwood_communication>=2.2.4
-pyarrow
 streamlit
+extra-streamlit-components
 streamlit-cookies-controller
+pandas
+lxml
+pyarrow
+pumpwood_communication>=2.2.4
+openpyxl
+loguru>=0.7.3
